1297 - move Layout to a seperate component
